backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
openai==1.3.0
langchain==0.0.200
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pydantic==1.10.0